move to cargo-fuzz
