+++ ci/common.sh
@@ -1,0 +1,1 @@
+TESTS="str_repeat vec_deque hash_map"
