+++ fuzz/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "bughunt-rust-fuzz"
+version = "0.0.1"
+authors = ["Automatically generated"]
+publish = false
+
+[package.metadata]
+cargo-fuzz = true
+
+[dependencies]
+arbitrary = "0.2"
+bh_alloc  = "0.2.4"
+[dependencies.bughunt-rust]
+path = ".."
+[dependencies.libfuzzer-sys]
+git = "https://github.com/rust-fuzz/libfuzzer-sys.git"
+
+# Prevent this from interfering with workspaces
+[workspace]
+members = ["."]
+
+[[bin]]
+path = "fuzz_targets/stdlib/str/repeat.rs"
+name = "str_repeat"
+
+[[bin]]
+path = "fuzz_targets/stdlib/collections/hash_map.rs"
+name = "hash_map"
+
+[[bin]]
+path = "fuzz_targets/stdlib/collections/vec_deque.rs"
+name = "vec_deque"
